[hevce] Remove legacy path (Phase 2)
Remaining dependencies on legacy path are removed for HEVCe
